Integrate MongoDB and add 2 XP commands
Integrate MongoDB to the bot. This connects the bot to MongoDB each time it boots and this will read the database for each time it's accessed (to get xp for the relevant user etc)

Added 2 commands - /xp to view the user's XP and /editxp to edit the amount of XP they have, both contacting the database.

Added a schema that structures how the XP is stored. This is most likely temporary and will be expanded on.