Add full equip functionality
Fully implement equip functionality for items that are in a user's inventory. A user needs to specify the item name and it will automatically equip to the badge or colour slot based on what it is.

Added early returns for some edge cases where users don't own items or item cannot be found